+++ hibikase/KaraokeData/TempoMap.h
@@ -5,6 +5,8 @@
 #include <chrono>
 #include <map>
 
+#include <QtGlobal>
+
 namespace KaraokeData
 {
 
@@ -13,9 +15,9 @@ class TempoMap
 public:
     TempoMap();
 
-    void AddTempoEntry(int start_ticks, std::chrono::nanoseconds tick_duration);
+    void AddTempoEntry(qint64 start_ticks, std::chrono::nanoseconds tick_duration);
 
-    std::chrono::nanoseconds TicksToTime(int ticks) const;
+    std::chrono::nanoseconds TicksToTime(qint64 ticks) const;
 
     bool empty() const { return m_tempo_map.empty(); }
 
@@ -24,13 +26,13 @@ private:
     {
         std::chrono::nanoseconds tick_duration = std::chrono::nanoseconds::zero();
         std::chrono::nanoseconds start_time = std::chrono::nanoseconds::zero();
-        int start_ticks = 0;
+        qint64 start_ticks = 0;
     };
 
-    static constexpr int MAX_TICK = std::numeric_limits<int>().max();
+    static constexpr qint64 MAX_TICK = std::numeric_limits<qint64>().max();
 
     // The key is the tick a given tempo entry ends at
-    std::map<int, TempoEntry> m_tempo_map;
+    std::map<qint64, TempoEntry> m_tempo_map;
 };
 
 } // namespace KaraokeData
